Some optimizations and readability

+++ server/src/lsp_types.rs
@@ -135,13 +135,13 @@ pub enum CompilerProgressParams {
     },
     Completed {
         success: bool,
-        code: isize,
+        code: i32,
         lines: Vec<String>,
     },
     SingleProjectCompleted {
         project_id: usize,
         success: bool,
-        code: isize,
+        code: i32,
         lines: Vec<String>,
     },
 }
@@ -170,7 +170,7 @@ impl CompilerProgress {
         }).await;
     }
 
-    pub async fn notify_completed(client: &tower_lsp::Client, success: bool, code: isize, lines: Vec<String>) {
+    pub async fn notify_completed(client: &tower_lsp::Client, success: bool, code: i32, lines: Vec<String>) {
         client.send_notification::<CompilerProgress>(CompilerProgressParams::Completed {
             success,
             code,
@@ -178,7 +178,13 @@ impl CompilerProgress {
         }).await;
     }
 
-    pub async fn notify_single_project_completed(client: &tower_lsp::Client, project_id: usize, success: bool, code: isize, lines: Vec<String>) {
+    pub async fn notify_single_project_completed(
+        client: &tower_lsp::Client,
+        project_id: usize,
+        success: bool,
+        code: i32,
+        lines: Vec<String>
+    ) {
         client.send_notification::<CompilerProgress>(CompilerProgressParams::SingleProjectCompleted {
             project_id,
             success,
